+++ envoy/network/dns.h
@@ -6,11 +6,11 @@
 #include <memory>
 #include <string>
 
-#include "envoy/common/optref.h"
 #include "envoy/common/pure.h"
 #include "envoy/common/time.h"
 #include "envoy/network/address.h"
 
+#include "absl/types/optional.h"
 #include "absl/types/variant.h"
 
 namespace Envoy {
@@ -56,7 +56,7 @@ public:
   virtual void addTrace(uint8_t trace) PURE;
 
   /** Return the DNS query traces. */
-  virtual OptRef<const std::vector<Trace>> getTraces() PURE;
+  virtual std::string getTraces() PURE;
 };
 
 /**