+++ source/common/orca/orca_parser.h
@@ -12,9 +12,9 @@ namespace Orca {
 static constexpr absl::string_view kEndpointLoadMetricsHeader = "endpoint-load-metrics";
 static constexpr absl::string_view kEndpointLoadMetricsHeaderBin = "endpoint-load-metrics-bin";
 // Prefix used to determine format expected in kEndpointLoadMetricsHeader.
-static constexpr absl::string_view kHeaderFormatPrefixBin = "BIN";
-static constexpr absl::string_view kHeaderFormatPrefixJson = "JSON";
-static constexpr absl::string_view kHeaderFormatPrefixText = "TEXT";
+static constexpr absl::string_view kHeaderFormatPrefixBin = "BIN ";
+static constexpr absl::string_view kHeaderFormatPrefixJson = "JSON ";
+static constexpr absl::string_view kHeaderFormatPrefixText = "TEXT ";
 // The following fields are the names of the metrics tracked in the ORCA load
 // report proto.
 static constexpr absl::string_view kApplicationUtilizationField = "application_utilization";