+++ source/extensions/clusters/dynamic_forward_proxy/cluster.cc
@@ -93,7 +93,9 @@ Cluster::~Cluster() {
     auto cluster_name = it->first;
     ENVOY_LOG(debug, "cluster='{}' removing from cluster_map & cluster manager", cluster_name);
     cluster_map_.erase(it++);
-    cm_.removeCluster(cluster_name);
+    cm_.removeCluster(cluster_name,
+                      Runtime::runtimeFeatureEnabled(
+                          "envoy.reloadable_features.avoid_dfp_cluster_removal_on_cds_update"));
   }
 }
 
@@ -131,7 +133,9 @@ void Cluster::checkIdleSubCluster() {
         auto cluster_name = it->first;
         ENVOY_LOG(debug, "cluster='{}' removing from cluster_map & cluster manager", cluster_name);
         cluster_map_.erase(it++);
-        cm_.removeCluster(cluster_name);
+        cm_.removeCluster(cluster_name,
+                          Runtime::runtimeFeatureEnabled(
+                              "envoy.reloadable_features.avoid_dfp_cluster_removal_on_cds_update"));
       } else {
         ++it;
       }