+++ test/mocks/upstream/cluster_manager.h
@@ -38,8 +38,8 @@ public:
   MOCK_METHOD(absl::Status, initialize, (const envoy::config::bootstrap::v3::Bootstrap& bootstrap));
   MOCK_METHOD(bool, initialized, ());
   MOCK_METHOD(bool, addOrUpdateCluster,
-              (const envoy::config::cluster::v3::Cluster& cluster,
-               const std::string& version_info));
+              (const envoy::config::cluster::v3::Cluster& cluster, const std::string& version_info,
+               const bool avoid_cds_removal));
   MOCK_METHOD(void, setPrimaryClustersInitializedCb, (PrimaryClustersReadyCallback));
   MOCK_METHOD(void, setInitializedCb, (InitializationCompleteCallback));
   MOCK_METHOD(absl::Status, initializeSecondaryClusters,
@@ -48,7 +48,7 @@ public:
 
   MOCK_METHOD(const ClusterSet&, primaryClusters, ());
   MOCK_METHOD(ThreadLocalCluster*, getThreadLocalCluster, (absl::string_view cluster));
-  MOCK_METHOD(bool, removeCluster, (const std::string& cluster));
+  MOCK_METHOD(bool, removeCluster, (const std::string& cluster, const bool remove_ignored));
   MOCK_METHOD(void, shutdown, ());
   MOCK_METHOD(bool, isShutdown, ());
   MOCK_METHOD(const absl::optional<envoy::config::core::v3::BindConfig>&, bindConfig, (), (const));